examples/requirements/django_2_0.txt
-r common.txt
-r testing.txt

Django>=2.0,<2.1
django-debug-toolbar>=1.6
sqlparse>=0.2.2
django-nine
